+++ idaes_ui/fv/models/tests/test_issues.py
@@ -2,13 +2,8 @@
 Tests for issues module
 """
 import pytest
-from idaes_ui.fv.tests.flowsheets import idaes_demo_flowsheet
 from ..issues import *
-
-
-@pytest.fixture
-def flowsheet():
-    return idaes_demo_flowsheet()
+from . import flowsheet
 
 
 @pytest.mark.unit
